pyproject.toml
[project]
name = "attendance-agent"
version = "1.0.0"
description = "MS Teams Auto-Join Attendance Agent - automatically joins your scheduled Teams classes"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
dependencies = [
    "playwright>=1.49,<2",
    "pyyaml>=6.0,<7",
    "plyer>=2.1,<3",
]

[project.scripts]
attendance-agent = "src.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
